Rename constants.js to constants/urls.js so it does not conflict node native constant package.json entry when webpack tries to resolve it (#8136)
