+++ _inc/client/components/welcome-new-plan/personal.jsx
@@ -11,7 +11,7 @@ import analytics from 'lib/analytics';
  */
 import JetpackDialogue from 'components/jetpack-dialogue';
 import Button from 'components/button';
-import { imagePath } from 'constants';
+import { imagePath } from 'constants/urls';
 
 class WelcomePersonal extends Component {
 	componentDidMount() {
